nest: Add wireless topology map
Added a topology map for wireless networks, which stores
a list of BSS's and IBSS's in the topology, their SSIDs,
and a list of stations that are part of them, and implemented
functions for updating and retrieving information from the topology map.
Using this, APIs have been created for a node to leave the wireless
network that it is part of. Also validation checks are done while
joining or leaving a network, using information in the topology map.
Also new example programs have been written to demonstrate the usage.

Signed-off-by: Niranjan S Yadiyala <niranjan.sy99@gmail.com>